requirements.txt
numpy==1.24.3
pandas==2.0.3
sqlalchemy==2.0.27
openai==1.12.0
streamlit==1.31.1
plotly==5.18.0
python-dotenv==1.0.1
